backend/pyproject.toml
[project]
name = "frame-mog"
version = "0.1.0"
description = "frame-mog FastAPI backend"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.136.1",
    "uvicorn[standard]>=0.46.0",
    "pydantic>=2.8",
    "pydantic-settings>=2.0",
    "pydantic-ai-slim[openai]>=1.87.0",
    "python-dotenv>=1.0.1",
    "python-multipart>=0.0.9",
    "piccolo>=1.0.0",
    "asyncpg>=0.29.0",
    "clerk-backend-api>=5.0.0",
]

[dependency-groups]
dev = [
    "httpx>=0.27",
    "pytest>=8",
    "ruff>=0.15.0",
]
